move `generateModal` to `View.hs`
