+++ src/Swarm/TUI/Controller.hs
@@ -61,10 +61,8 @@ import Witch (into)
 import Brick hiding (Direction)
 import Brick.Focus
 import Brick.Forms
-import Brick.Widgets.Center (hCenter)
 import Brick.Widgets.Dialog
 import qualified Brick.Widgets.List as BL
-import qualified Brick.Widgets.Table as BT
 import qualified Graphics.Vty as V
 
 import qualified Control.Carrier.Lift as Fused
@@ -84,6 +82,7 @@ import Swarm.Language.Syntax
 import Swarm.Language.Types
 import Swarm.TUI.List
 import Swarm.TUI.Model
+import Swarm.TUI.View (generateModal)
 import Swarm.Util hiding ((<<.=))
 
 -- | Pattern synonyms to simplify brick event handler
@@ -151,55 +150,6 @@ toggleModal s mt = do
   -- TODO: manage unpause more safely to also cover
   -- the world event handler for the KChar 'p'.
   resetLastFrameTime curTime = uiState . lastFrameTime .~ curTime
-
-generateModal :: ModalType -> Modal
-generateModal mt = Modal mt (dialog (Just title) buttons (maxModalWindowWidth `min` requiredWidth)) widget
- where
-  (title, widget, buttons, requiredWidth) =
-    case mt of
-      HelpModal -> ("Help", helpWidget, Nothing, maxModalWindowWidth)
-      WinModal -> ("", txt "Congratulations!", Nothing, maxModalWindowWidth)
-      QuitModal ->
-        let quitMsg = "Are you sure you want to quit?"
-         in ( ""
-            , padBottom (Pad 1) $ hCenter $ txt quitMsg
-            , Just (0, [("Cancel", False), ("Quit", True)])
-            , T.length quitMsg + 4
-            )
-
-helpWidget :: Widget Name
-helpWidget = (helpKeys <=> fill ' ') <+> (helpCommands <=> fill ' ')
- where
-  helpKeys =
-    vBox
-      [ hCenter $ txt "Global Keybindings"
-      , hCenter $ mkTable glKeyBindings
-      ]
-  mkTable = BT.renderTable . BT.table . map toWidgets
-  toWidgets (k, v) = [txt k, txt v]
-  glKeyBindings =
-    [ ("F1", "Help")
-    , ("Ctrl-q", "quit the game")
-    , ("Tab", "cycle panel focus")
-    , ("Meta-w", "focus on the world map")
-    , ("Meta-e", "focus on the robot inventory")
-    , ("Meta-r", "focus on the REPL")
-    , ("Meta-t", "focus on the info panel")
-    ]
-  helpCommands =
-    vBox
-      [ hCenter $ txt "Commands"
-      , hCenter $ mkTable baseCommands
-      ]
-  baseCommands =
-    [ ("build {<commands>}", "Create a robot")
-    , ("make \"<name>\"", "Craft an item")
-    , ("move", "Move one step in the current direction")
-    , ("turn <dir>", "Change the current direction")
-    , ("grab", "Grab whatver is available")
-    , ("give <robot> \"<item>\"", "Give an item to another robot")
-    , ("has \"<item>\"", "Check for an item in the inventory")
-    ]
 
 handleModalEvent :: AppState -> V.Event -> EventM Name (Next AppState)
 handleModalEvent s ev = case ev of
